revert change,
The build explicitly limits to Firefox as this is used for the selenium tests. Not including the others has the advantage that the build runs 5 times faster. This project is not intended to be distributed to others, the gwt project under "application" serves that purpose, so I don't think there is any advantage in compiling for all browsers.